Created tables for anime and users and corresponding JPA classes

+++ backend/src/main/resources/db/migration/V2__Create_anime_list_entries_table.sql
@@ -2,8 +2,8 @@
 -- This table stores each anime in a user's personal list
 
 CREATE TABLE anime_list_entries (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL,
     anilist_id INTEGER NOT NULL,
     status VARCHAR(20) NOT NULL DEFAULT 'plan_to_watch',
     score INTEGER CHECK (score >= 1 AND score <= 10),
